Move card controls to component
