+++ src/components/card-controls.js
@@ -1,0 +1,42 @@
+import AbstractComponent from './abstract-component';
+
+import {getClass, getFilmControlsData} from '../helpers';
+
+export default class CardControls extends AbstractComponent {
+  constructor({isInWatchList, isWatched, isFavorite}) {
+    super();
+
+    this._controlsData = getFilmControlsData({
+      isInWatchList,
+      isWatched,
+      isFavorite,
+    });
+  }
+  _getCardControl({id, text, isActive}) {
+    const mods = [id];
+
+    if (isActive) {
+      mods.push(`active`);
+    }
+
+    const className = getClass({
+      base: `film-card__controls-item`,
+      mods
+    });
+
+    return (
+      `<button class="${className}">${text}</button>`
+    );
+  }
+
+  _getTmpl() {
+    const controlsMarkup = this._controlsData
+      .reduce((prev, control) => prev + this._getCardControl(control), ``);
+
+    return (
+      `<form class="film-card__controls">
+        ${controlsMarkup}
+      </form>`
+    );
+  }
+}
